backend/requirements/base.txt
# requirements/base.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.1
django-allauth>=0.59.0
social-auth-app-django>=5.4.0
PyJWT>=2.8.0
requests>=2.31.0

# Autocomplete
django-autocomplete-light>=3.9.7
django-import-export>=3.3.1

# API Documentation
drf-spectacular>=0.27.0

# Utils
python-decouple>=3.8
Pillow>=10.1.0
django-storages>=1.14.2
boto3>=1.34.0

# Celery (Background Tasks)
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# Utilities
bleach>=6.1.0
python-slugify>=8.0.1
django-extensions>=3.2.3

# Text-to-Speech (Neural TTS for pronunciation)
edge-tts>=6.1.0
aiofiles>=23.2.1

# Audio Processing
pydub>=0.25.1
mutagen>=1.47.0

# Speech-to-Text (Phase 5: Pronunciation Assessment)
google-cloud-speech>=2.21.0
